[FIX] Перенести запрет на удаление отдела в слой сервиса

+++ src/main/java/project/service/SubdepartmentServiceImpl.java
@@ -21,6 +21,7 @@ import static project.dataSource.DynamicDataSourceContextHolder.setCurrentDataSo
 public class SubdepartmentServiceImpl {
 
     private final SubdepartmentJpaRepository repository;
+    private final StudentServiceImpl studentServiceImpl;
 
     public @NonNull List<Subdepartment> findAll(int departmentId) {
         setCurrentDataSource("DEP_" + departmentId);
@@ -34,8 +35,10 @@ public class SubdepartmentServiceImpl {
     public boolean deleteById(int departmentId, short id) {
         try {
             setCurrentDataSource("DEP_" + departmentId);
+            if (studentServiceImpl.getStudentExistsWithSubdepartmentId(departmentId, id)) {
+                throw new IllegalArgumentException("Существует работник из этого отдела. Необходимо добавить его в другой отдел.");
+            }
             repository.deleteById(id);
-
             return true;
         } catch (EmptyResultDataAccessException e) {
             return false;
